move eval results and actions
